[Storage management] Change spaced DBus response type
Since crrev.com/c/3260777, spaced returns int64_t instead of uint64_t.
Fix the spaced client and the storage page accordingly.

Bug: b:177966496, b:173086361
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291259
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944836}

+++ chrome/browser/ui/webui/settings/chromeos/calculator/size_calculator.cc
@@ -103,9 +103,9 @@ void TotalDiskSpaceCalculator::GetRootDeviceSize() {
 }
 
 void TotalDiskSpaceCalculator::OnGetRootDeviceSize(
-    absl::optional<uint64_t> reply) {
+    absl::optional<int64_t> reply) {
   if (reply.has_value()) {
-    NotifySizeCalculated(static_cast<int64_t>(reply.value()));
+    NotifySizeCalculated(reply.value());
     return;
   }
